fix(core): fix non-partitioned table truncate error
