+++ core/src/test/java/io/questdb/test/griffin/AbstractO3Test.java
@@ -26,13 +26,16 @@ package io.questdb.test.griffin;
 
 import io.questdb.Metrics;
 import io.questdb.cairo.*;
-import io.questdb.griffin.SqlCompiler;
-import io.questdb.griffin.SqlException;
-import io.questdb.griffin.SqlExecutionContext;
+import io.questdb.cairo.sql.Record;
+import io.questdb.cairo.sql.RecordCursor;
+import io.questdb.cairo.sql.RecordCursorFactory;
+import io.questdb.cairo.sql.RecordMetadata;
+import io.questdb.griffin.*;
 import io.questdb.griffin.engine.functions.rnd.SharedRandom;
 import io.questdb.log.Log;
 import io.questdb.log.LogFactory;
 import io.questdb.mp.WorkerPool;
+import io.questdb.std.BytecodeAssembler;
 import io.questdb.std.FilesFacade;
 import io.questdb.std.Rnd;
 import io.questdb.test.std.TestFilesFacadeImpl;
@@ -53,6 +56,7 @@ import java.net.URL;
 import java.util.concurrent.TimeUnit;
 
 public class AbstractO3Test {
+
     protected static final StringSink sink = new StringSink();
     protected static final StringSink sink2 = new StringSink();
     private final static Log LOG = LogFactory.getLog(O3Test.class);
@@ -66,6 +70,7 @@ public class AbstractO3Test {
             .withTimeout(20 * 60 * 1000, TimeUnit.MILLISECONDS)
             .withLookingForStuckThread(true)
             .build();
+    private RecordToRowCopier copier;
 
     @BeforeClass
     public static void setupStatic() {
@@ -74,6 +79,11 @@ public class AbstractO3Test {
         } catch (IOException e) {
             throw new RuntimeException();
         }
+    }
+
+    @Before
+    public void clearRecordToRowCopier() {
+        copier = null;
     }
 
     @Before
@@ -279,6 +289,22 @@ public class AbstractO3Test {
         assertMaxTimestamp(compiler.getEngine(), compiler, sqlExecutionContext, "select max(ts) from y");
     }
 
+    protected static void assertXY(SqlCompiler compiler, SqlExecutionContext sqlExecutionContext) throws SqlException {
+        TestUtils.assertEquals(
+                compiler,
+                sqlExecutionContext,
+                "select * from x",
+                "select * from y"
+        );
+
+        TestUtils.assertEquals(
+                compiler,
+                sqlExecutionContext,
+                "select count() from x",
+                "select count() from y"
+        );
+    }
+
     static void executeVanilla(TestUtils.LeakProneCode code) throws Exception {
         AbstractO3Test.assertMemoryLeak(code);
     }
@@ -403,5 +429,36 @@ public class AbstractO3Test {
             String sql
     ) throws SqlException {
         TestUtils.printSql(compiler, sqlExecutionContext, sql, sink);
+    }
+
+    protected void insertUncommitted(
+            SqlCompiler compiler,
+            SqlExecutionContext sqlExecutionContext,
+            String sql,
+            TableWriter writer
+    ) throws SqlException {
+        try (RecordCursorFactory factory = compiler.compile(sql, sqlExecutionContext).getRecordCursorFactory()) {
+            RecordMetadata metadata = factory.getMetadata();
+            int timestampIndex = writer.getMetadata().getTimestampIndex();
+            EntityColumnFilter toColumnFilter = new EntityColumnFilter();
+            toColumnFilter.of(metadata.getColumnCount());
+            if (null == copier) {
+                copier = RecordToRowCopierUtils.generateCopier(
+                        new BytecodeAssembler(),
+                        metadata,
+                        writer.getMetadata(),
+                        toColumnFilter
+                );
+            }
+            try (RecordCursor cursor = factory.getCursor(sqlExecutionContext)) {
+                final Record record = cursor.getRecord();
+                while (cursor.hasNext()) {
+                    long timestamp = record.getTimestamp(timestampIndex);
+                    TableWriter.Row row = writer.newRow(timestamp);
+                    copier.copy(record, row);
+                    row.append();
+                }
+            }
+        }
     }
 }
